Added adverts table

+++ src/main/java/com/piciu1221/starmoto/repository/CarRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {
 
     Optional<Object> findByVin(String vin);
